requirements.txt
celery==5.0.5
requests==2.25.1
PyYAML==5.4.1
